@@ -2,15 +2,10 @@
   <div id="#app" class="flex flex-col min-h-screen">
     <router-view class="mx-auto my-auto p-8"></router-view>
     <footer class="mt-auto mb-1 text-center text-gray-200">
-      <p>Cosme Gressier | @ DevChallenges.io | <a href="https://github.com/Cosme711">@Cosme711</a></p>
+      <p>
+        Cosme Gressier | @ DevChallenges.io |
+        <a href="https://github.com/Cosme711">@Cosme711</a>
+      </p>
     </footer>
   </div>
 </template>
-
-<script>
-
-export default {
-  name: 'App'
-}
-</script>
-
